// frontend/sysbimbo-app/src/app/core/models/sku-catalogo.model.ts
export interface SkuCatalogo {
  codigoSkuBimbo: string;
  codigoSkuB2B: string | null;
  nombreSkuBimbo: string | null;
  nombreSkuB2B: string | null;
  unidadNegocio: string | null;
  area: string | null;
  categoria: string | null;
  marca: string | null;
  tipoProducto: string | null;
  status: string | null;
  gramaje: string | null;
}